requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==2.0
djangorestframework==3.8.2
get==1.0.3
idna==2.7
mysqlclient==1.3.12
post==1.0.2
psycopg2-binary==2.7.4
public==1.0.3
pytz==2018.4
query-string==1.0.2
requests==2.19.1
urllib3==1.23
python-magic
